+++ service/controller/resource/servicemonitor/resource.go
@@ -54,6 +54,7 @@ func toServiceMonitors(obj interface{}) ([]*promv1.ServiceMonitor, error) {
 
 	return []*promv1.ServiceMonitor{
 		service.APIServer(cluster),
+		service.NginxIngressController(cluster),
 	}, nil
 }